Add a couple of special cases for string literals
We want some string literals to act differently, because they can end up
in attribute-like places. So far I've identified those as:
 - Component attributes (`componentAttrs`)
 - Positional parameters in block components
   (`blockPositionalParameters`)

If not specified, those default to the value of `attrs`.

This also checks that the attributes are contained by a component - this
means that element modifier statements do not count as component
attributes, so if string literals are set to single quote and attributes
are set to double, the following is correct per the linter:

<form {{action 'foo' on='bar'}}>

This may be too finicky, and I might change my mind on separating those
out from all hash values.